move some functions to utils instead of have them in Types.jl
