+++ src/REPLMode/argument_parsers.jl
@@ -1,3 +1,5 @@
+import ..isdir_windows_workaround
+
 ###############
 # PackageSpec #
 ###############
@@ -106,7 +108,7 @@ end
 function parse_registry(word::AbstractString; add=false)::RegistrySpec
     word = replace(word, "~" => homedir())
     registry = RegistrySpec()
-    if add && Types.isdir_windows_workaround(word) # TODO: Should be casesensitive_isdir
+    if add && isdir_windows_workaround(word) # TODO: Should be casesensitive_isdir
         if isdir(joinpath(word, ".git")) # add path as url and clone it from there
             registry.url = abspath(word)
         else # put the path
