+++ src/utils.jl
@@ -12,3 +12,20 @@ function set_readonly(path)
     return nothing
 end
 
+# try to call realpath on as much as possible
+function safe_realpath(path)
+    ispath(path) && return realpath(path)
+    a, b = splitdir(path)
+    return joinpath(safe_realpath(a), b)
+end
+
+# Windows sometimes throw on `isdir`...
+function isdir_windows_workaround(path::String)
+    try isdir(path)
+    catch e
+        false
+    end
+end
+
+casesensitive_isdir(dir::String) =
+    isdir_windows_workaround(dir) && basename(dir) in readdir(joinpath(dir, ".."))
