+++ test/pkg.jl
@@ -300,7 +300,7 @@ temp_pkg_dir() do project_path
 
     @testset "check logging" begin
         usage = Pkg.TOML.parse(String(read(joinpath(Pkg.logdir(), "manifest_usage.toml"))))
-        manifest = Types.safe_realpath(joinpath(project_path, "Manifest.toml"))
+        manifest = Pkg.safe_realpath(joinpath(project_path, "Manifest.toml"))
         @show keys(usage)
         @show manifest
         @show usage
